cambien sueldo hora de int a double

+++ RegistroTecnico/Models/Tecnicos.cs
@@ -9,6 +9,6 @@ namespace RegistroTecnico.Models
 
         [Required(ErrorMessage = "este campo es obligatorio")]
         public string Nombre { get; set; } = string.Empty!; // para que nombre no sea nulo
-        public int SueldoHora { get; set; }
+        public double SueldoHora { get; set; }
     }
 }
